feat: update post routes to associate posts with authenticated user and integrate react-router

+++ backend/src/services/posts.js
@@ -1,7 +1,7 @@
 import { Post } from '../db/models/post.js'
 
-export async function createPost({ title, author, contents, tags }) {
-  const post = new Post({ title, author, contents, tags })
+export async function createPost(userID, { title, contents, tags }) {
+  const post = new Post({ title, author: userID, contents, tags })
   return await post.save()
 }
 
@@ -24,13 +24,13 @@ export async function listPostsByTag(tags, options) {
 export async function getPostById(postId) {
   return await Post.findById(postId)
 }
-export async function updatePost(postId, { title, author, contents, tags }) {
+export async function updatePost(userId, postId, { title, contents, tags }) {
   return await Post.findOneAndUpdate(
-    { _id: postId },
-    { $set: { title, author, contents, tags } },
+    { _id: postId, author: userId },
+    { $set: { title, contents, tags } },
     { new: true },
   )
 }
-export async function deletePost(postId) {
-  return await Post.deleteOne({ _id: postId })
+export async function deletePost(userId, postId) {
+  return await Post.deleteOne({ _id: postId, author: userId })
 }
